requirements.txt
PyQt5==5.15.9
requests==2.31.0
psutil==5.9.5
pywin32==306; sys_platform == "win32"
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
yt-dlp==2023.11.16
Pillow==10.1.0 